+++ src/model/resolver/amazon-details.ts
