+++ src/model/resolver/audible-information.ts
