+++ src/model/resolver/scribd-information.ts
