+++ src/service/resolver/audible/audible-resolver.service.ts
@@ -2,12 +2,12 @@ import { HTMLElement } from 'node-html-parser';
 import { URL } from 'url';
 
 import { NullableHtmlElement } from '../../../model/html/nullable-html-element';
+import { AudibleAuthor, AudibleInformation } from '../../../model/resolver/audible-information';
 import { SiteResolver } from '../../../model/resolver/site-resolver.enum';
 import { Message } from '../../../model/telegram/message';
 import { HtmlUtil } from '../../../util/html-util';
 import { AbstractResolver } from '../abstract-resolver';
 import { StatisticsService } from './../../statistics/statistic.service';
-import { AudibleAuthor, AudibleInformation } from './audible-information';
 
 export class AudibleResolverService extends AbstractResolver {
   private static readonly OVERRIDE_LANGUAGE = 'ipRedirectOverride';
